service renamed: redis -> redis_tasks_queue, added redis_statistics container, fixed java worker not building
added pack of 100 flash samples, speed up flash decompiler

split flash_pack.zip

squashed commits (file from older commit was > 100mb)

added tests for flash decompilation ratio, new test case
