+++ .circleci/config.yml
@@ -8,18 +8,18 @@ jobs:
     steps:
       - checkout
       - run: free -h
-      - run: pip install --retries 10 --timeout 30 docker-compose==1.20.1 docker==3.1.4
+      - run: pip install --retries 10 --timeout 30 docker-compose==1.24.1 docker==4.0.2
       - run: docker-compose up -d syslog
       - run: docker-compose up -d db
       - run:
-          command: docker-compose up -d api
+          command: docker-compose up -d
           no_output_timeout: 20m
       - run: sleep 10 && free -h && docker-compose ps
       - run:
           name: check containers. If at least one is down, print logs
           command: |
             if [ $(docker-compose  ps | grep exit | wc -l) -ne "0" ]; then
-               docker-compose exec syslog cat /var/log/messages
+               docker-compose exec syslog cat /var/log/messages;
             fi
       - run: docker-compose exec syslog cat /var/log/messages
       - run: docker-compose exec api sh -c "pip freeze"
@@ -28,18 +28,25 @@ jobs:
       - run: docker-compose exec api sh -c "python /daas/daas/manage.py migrate"
       - run:
           name: executing tests
-          command: docker-compose exec api sh -c "coverage run --source='/daas/daas/daas_app' /daas/daas/manage.py test daas_app.tests -v 1 --force-color --exclude stress_test_csharp --noinput "
+          command: docker-compose exec api sh -c "coverage run --source='/daas/daas/daas_app' /daas/daas/manage.py test daas_app.tests.unit_tests -v 1 --force-color --exclude stress_test_csharp --noinput "
       - run:
           name: executing tests in reverse order
-          command: docker-compose exec api sh -c "python /daas/daas/manage.py test daas_app.tests --reverse -v 1 --force-color --exclude stress_test_csharp --noinput "
-      - run: docker-compose up -d redis
+          command: docker-compose exec api sh -c "python /daas/daas/manage.py test daas_app.tests.unit_tests --reverse -v 1 --force-color --exclude stress_test_csharp --noinput "
+      - run:
+          name: print logs on fail after tests
+          command: docker-compose exec syslog cat /var/log/messages
+          when: on_fail
       - run:
           name: executing stress tests
-          no_output_timeout: 240m
-          command: docker-compose exec api sh -c "python /daas/daas/manage.py test daas_app.tests.stress_test_csharp -v 1 --force-color --noinput "
+          no_output_timeout: 60m
+          command: |
+            if [[ ! -z $CI_PULL_REQUEST ]]; then
+               docker-compose exec api sh -c "python /daas/daas/manage.py test daas_app.tests.decompilation_ratio_tests -v 1 --force-color --noinput ";
+            fi
       - run:
-          name: check containers
-          command: docker-compose ps
+          name: print logs on fail after stress tests
+          command: docker-compose exec syslog cat /var/log/messages
+          when: on_fail
       - run:
           name: executing coverage
           command: docker-compose exec api sh -c "coverage report -m --fail-under=70 --omit='*daas_app/tests/*,*daas_app/migrations/*,*daas_app/decompilers*'"
