+++ Dockerfile
@@ -1,10 +1,9 @@
-FROM python:3.7.0-stretch
+FROM python:3.7.4-stretch
 RUN mkdir /daas
 WORKDIR /daas
-ADD . /daas
 ENV PYTHONUNBUFFERED=0
 ENV HOME /home/root
-RUN pip install --upgrade pip
+
 
 RUN apt-get clean && \
 apt-get update && \
@@ -12,4 +11,9 @@ apt-get install --no-install-recommends -y build-essential apt-transport-https &
 apt-get clean && \
 apt-get update
 
-RUN pip install -r /daas/pip_requirements_api.txt
+ADD . /daas
+RUN pip install --upgrade pip && \
+    pip install -r /daas/pip_requirements_api.txt && \
+    echo "Joining test resources..." && \
+    cat /daas/daas/daas/daas_app/tests/resources/flash_pack_parts/flash_pack.zip.? > \
+        /daas/daas/daas/daas_app/tests/resources/flash_pack_parts/flash_pack.zip
