+++ daas/daas_app/decompilers/worker.py
@@ -1,11 +1,12 @@
 from .decompiler_factory import DecompilerFactory
-from .utils import send_result, get_sample
+from .utils import DaaSAPIConnector
 
 
 # This function should by called by redis queue (rq command).
 def worker(task):
     decompiler = DecompilerFactory().create(task['config'])
     sample_id = task['sample_id']
-    sample = get_sample(sample_id)
+    connector = DaaSAPIConnector(task['api_base_url'])
+    sample = connector.get_sample(sample_id)
     result = decompiler.process(sample)
-    send_result(result)
+    connector.send_result(result)
