+++ daas/daas_app/tests/decompilation_ratio_tests/test_flash_decompilation_ratio.py
@@ -1,0 +1,13 @@
+from ..test_utils.test_cases.decompilation_ratio import DecompilationRatioTestCase
+from ..test_utils.resource_directories import FLASH_ZIPPED_PACK
+
+
+class CsharpTest(DecompilationRatioTestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.zipped_samples_path = FLASH_ZIPPED_PACK
+        cls.timeout_per_sample = 1200
+        cls.decompiled = 100
+        cls.timed_out = 0
+        cls.failed, = 0
+        super().setUpClass()
