+++ daas/daas_app/tests/resources/password.MD
@@ -1,0 +1,1 @@
+The password is "codex"
