+++ daas/daas_app/tests/test_utils/resource_directories.py
@@ -1,0 +1,7 @@
+CSHARP_SAMPLE = '/daas/daas/daas_app/tests/resources/460f0c273d1dc133ed7ac1e24049ac30.csharp'
+CSHARP_ZIPPED_PACK = '/daas/daas/daas_app/tests/resources/csharp_pack.zip'
+TEXT_SAMPLE = '/daas/daas/daas_app/tests/resources/text.txt'
+FLASH_SAMPLE_01 = '/daas/daas/daas_app/tests/resources/995bb44df3d6b31d9422ddb9f3f78b7b.flash'
+FLASH_SAMPLE_02 = '/daas/daas/daas_app/tests/resources/eb19009c086845d0408c52d495187380c5762b8c.flash'
+FLASH_ZIPPED_PACK = '/daas/daas/daas_app/tests/resources/flash_pack_parts/flash_pack.zip'
+ZIP = '/daas/daas/daas_app/tests/resources/zip.zip'
