+++ daas/daas_app/utils/connections/django_server.py
@@ -1,0 +1,17 @@
+import socket
+from django.conf import settings
+import sys
+
+
+class DjangoServerConfiguration:
+    @property
+    def ip(self):
+        return socket.gethostbyname(socket.gethostname())
+
+    @property
+    def port(self):
+        return settings.TEST_INSTANCE_PORT if ('test' in sys.argv) else settings.DEFAULT_PORT
+
+    @property
+    def base_url(self):
+        return f'{self.ip}:{self.port}'
