+++ javaWorkerDockerfile
@@ -1,95 +1,41 @@
-FROM python:3.7.0-stretch
+FROM python:3.7.4-buster
 RUN mkdir /daas
 WORKDIR /daas
-ADD . /daas
 ENV PYTHONUNBUFFERED=0
 ENV HOME /home/root
 
 
 # Generic
-RUN apt-get clean && \
-apt-get update && \
-apt-get install --no-install-recommends -y build-essential apt-transport-https && \
-apt install --assume-yes gnutls-bin zlibc zlib1g zlib1g-dev \
-unzip \
-xauth \
-zenity \
-xvfb \
-host
-
-
-# Added for java decompiler:
-
-
-###### start of code from https://github.com/docker-library/openjdk/blob/master/11/jdk/Dockerfile ######
-RUN apt-get update && apt-get install -y --no-install-recommends \
-		bzip2 \
-		unzip \
-		xz-utils \
-	&& rm -rf /var/lib/apt/lists/*
-
-RUN echo 'deb http://deb.debian.org/debian stretch-backports main' > /etc/apt/sources.list.d/stretch-backports.list
-
-# Default to UTF-8 file.encoding
-ENV LANG C.UTF-8
-
-# add a simple script that can auto-detect the appropriate JAVA_HOME value
-# based on whether the JDK or only the JRE is installed
-RUN { \
-		echo '#!/bin/sh'; \
-		echo 'set -e'; \
-		echo; \
-		echo 'dirname "$(dirname "$(readlink -f "$(which javac || which java)")")"'; \
-	} > /usr/local/bin/docker-java-home \
-	&& chmod +x /usr/local/bin/docker-java-home
-
-# do some fancy footwork to create a JAVA_HOME that's cross-architecture-safe
-RUN ln -svT "/usr/lib/jvm/java-11-openjdk-$(dpkg --print-architecture)" /docker-java-home
-ENV JAVA_HOME /docker-java-home
-
-ENV JAVA_VERSION 11.0.3
-ENV JAVA_DEBIAN_VERSION 11.0.3+1-1~bpo9+2
-
-RUN set -ex; \
-	\
-# deal with slim variants not having man page directories (which causes "update-alternatives" to fail)
-	if [ ! -d /usr/share/man/man1 ]; then \
-		mkdir -p /usr/share/man/man1; \
-	fi; \
-	\
-# ca-certificates-java does not work on src:openjdk-11 with no-install-recommends: (https://bugs.debian.org/914860, https://bugs.debian.org/775775)
-# /var/lib/dpkg/info/ca-certificates-java.postinst: line 56: java: command not found
-	ln -svT /docker-java-home/bin/java /usr/local/bin/java; \
-	\
-	apt-get update; \
-	apt-get install -y --no-install-recommends \
-		openjdk-11-jdk="$JAVA_DEBIAN_VERSION" \
-	; \
-	rm -rf /var/lib/apt/lists/*; \
-	\
-	rm -v /usr/local/bin/java; \
-	\
-# ca-certificates-java does not work on src:openjdk-11: (https://bugs.debian.org/914424, https://bugs.debian.org/894979, https://salsa.debian.org/java-team/ca-certificates-java/commit/813b8c4973e6c4bb273d5d02f8d4e0aa0b226c50#d4b95d176f05e34cd0b718357c532dc5a6d66cd7_54_56)
-	keytool -importkeystore -srckeystore /etc/ssl/certs/java/cacerts -destkeystore /etc/ssl/certs/java/cacerts.jks -deststoretype JKS -srcstorepass changeit -deststorepass changeit -noprompt; \
-	mv /etc/ssl/certs/java/cacerts.jks /etc/ssl/certs/java/cacerts; \
-	/var/lib/dpkg/info/ca-certificates-java.postinst configure; \
-	\
-# verify that "docker-java-home" returns what we expect
-	[ "$(readlink -f "$JAVA_HOME")" = "$(docker-java-home)" ]; \
-	\
-# update-alternatives so that future installs of other OpenJDK versions do not change /usr/bin/java
-	update-alternatives --get-selections | awk -v home="$(readlink -f "$JAVA_HOME")" 'index($3, home) == 1 { $2 = "manual"; print | "update-alternatives --set-selections" }'; \
-# ... and verify that it actually worked for one of the alternatives we care about
-	update-alternatives --query java | grep -q 'Status: manual'
-###### end of code from https://github.com/docker-library/openjdk/blob/master/11/jdk/Dockerfile ######
+RUN apt-get update && \
+    apt-get install --no-install-recommends -y build-essential apt-transport-https && \
+    apt-get install --no-install-recommends -y gnutls-bin \
+        host \
+        unzip \
+        xauth \
+        xvfb \
+        zenity \
+        zlib1g \
+        zlib1g-dev \
+        zlibc && \
+    rm -rf /var/lib/apt/lists/* && \
+    apt-get clean
 
 
 
+# Install OpenJDK-11
+RUN apt update && \
+    echo 'deb http://ftp.de.debian.org/debian buster main' >> /etc/apt/sources.list && \
+    apt update && \
+    apt-get install --no-install-recommends -y openjdk-11-jdk && \
+    rm -rf /var/lib/apt/lists/* && \
+    apt-get clean
 
-## Downloading CFR:
+
+
+# Downloading CFR:
 RUN echo "installing cfr..." && \
-mkdir /cfr && \
-cd /cfr && \
-wget -nc https://www.benf.org/other/cfr/cfr-0.142.jar && \
-cd ../ && \
-echo "cfr installed"
+    mkdir /cfr && \
+    cd /cfr && \
+    wget -nc https://www.benf.org/other/cfr/cfr-0.142.jar && \
+    cd ../ && \
+    echo "cfr installed"
