+++ peWorkerDockerfile
@@ -1,7 +1,6 @@
-FROM python:3.7.0-stretch
+FROM python:3.7.4-stretch
 RUN mkdir /daas
 WORKDIR /daas
-ADD . /daas
 ENV PYTHONUNBUFFERED=0
 ENV HOME /home/root
 
@@ -11,46 +10,48 @@ ENV WINEARCH win32
 
 
 # Generic
-RUN apt-get clean && \
-apt-get update && \
-apt-get install --no-install-recommends -y build-essential apt-transport-https && \
-apt install --assume-yes gnutls-bin zlibc zlib1g zlib1g-dev \
-unzip \
-xauth \
-zenity \
-xvfb \
-host
+RUN apt-get update && \
+    apt-get install --no-install-recommends -y build-essential apt-transport-https && \
+    apt-get install --no-install-recommends -y gnutls-bin \
+        host \
+        unzip \
+        xauth \
+        xvfb \
+        zenity \
+        zlib1g \
+        zlib1g-dev \
+        zlibc && \
+    rm -rf /var/lib/apt/lists/* && \
+    apt-get clean
 
 
 # C#: Wine and wine's utils.
 RUN dpkg --add-architecture i386 && \
-apt-get clean && \
-apt-get update && \
-wget -nc https://dl.winehq.org/wine-builds/Release.key && \
-apt-key add Release.key && \
-echo "deb https://dl.winehq.org/wine-builds/debian/ stretch main" >> /etc/apt/sources.list.d/wine.list && \
-apt-get clean && \
-apt-get update && \
-apt install --assume-yes --allow-unauthenticated winehq-devel:i386=3.21.0~stretch \
-wine-devel:i386=3.21.0~stretch \
-wine-devel-i386:i386=3.21.0~stretch \
-fonts-wine \
-cabextract && \
-apt-get clean && \
-apt-get update
+    apt-get update && \
+    wget -nc https://dl.winehq.org/wine-builds/Release.key && \
+    apt-key add Release.key && \
+    echo "deb https://dl.winehq.org/wine-builds/debian/ stretch main" >> /etc/apt/sources.list.d/wine.list && \
+    apt-get update && \
+    apt install --assume-yes --allow-unauthenticated winehq-devel:i386=3.21.0~stretch \
+    wine-devel:i386=3.21.0~stretch \
+    wine-devel-i386:i386=3.21.0~stretch \
+    fonts-wine \
+    cabextract && \
+    rm -rf /var/lib/apt/lists/* && \
+    apt-get clean
 
 
 # C#: Winetricks, dotnet45, vcrun2010
 RUN echo "Installing winetricks" && \
-wget -nc -nv https://github.com/Winetricks/winetricks/archive/20181203.zip -O /tmp/winetricks.zip && \
-unzip /tmp/winetricks.zip -d /tmp/winetricks/ && \
-make -C /tmp/winetricks/winetricks-20181203 install && \
-rm -rf /tmp/winetricks.zip && \
-rm -rf /tmp/winetricks && \
-echo "Winetricks installed" && \
-echo "Installing Dotnet45"
-# Dot not use "&&" here because it will not work.
+    wget -nc -nv https://github.com/Winetricks/winetricks/archive/20181203.zip -O /tmp/winetricks.zip && \
+    unzip /tmp/winetricks.zip -d /tmp/winetricks/ && \
+    make -C /tmp/winetricks/winetricks-20181203 install && \
+    rm -rf /tmp/winetricks.zip && \
+    rm -rf /tmp/winetricks && \
+    echo "Winetricks installed" && \
+    echo "Installing Dotnet45"
+# Do not use "&&" here because it will not work.
 RUN timeout 1500 winetricks -q dotnet45 corefonts; if [ $? -eq 124 ]; then echo "Status is 124. Retrying dotnet45 installation..." && date && timeout 1000 winetricks -q dotnet45 corefonts; if [ $? -eq 124 ]; then "Status is 124 again(!). Retrying dotnet45 installation without timeout..." && date && winetricks -q dotnet45 corefonts; else echo "Status is not 124"; fi; else echo "Status is not 124."; fi; \
-echo "Dotnet45 installed"
-RUN pip install -r /daas/pip_requirements_worker.txt
-RUN xvfb-run winetricks -q vcrun2010 && echo "vcrun2010 installed"
+    echo "Dotnet45 installed"
+RUN xvfb-run winetricks -q vcrun2010 && \
+    echo "vcrun2010 installed"
